requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
web3>=7.0.0
cryptography==41.0.7
python-dotenv==1.0.0
django-cors-headers==4.3.1
pytz==2023.3

